+++ src/client/components/Search/Search.test.tsx
@@ -8,7 +8,7 @@ import { afterEach, beforeEach, describe, expect, test } from 'vitest';
 import * as bagResponse from './bag-response.json';
 import { Search } from './Search';
 import * as remoteConfig from './search-config.json';
-import { bffApi, remoteApi } from '../../../testing/test-utils';
+import { bffApi, remoteApi } from '../../../testing/utils';
 import { appStateAtom } from '../../hooks/useAppState';
 
 describe('<Search />', () => {
