+++ src/client/pages/Afis/Afis.test.tsx
@@ -6,7 +6,7 @@ import {
   AfisBusinessPartnerDetailsTransformed,
   AfisFacturenByStateResponse,
 } from '../../../server/services/afis/afis-types';
-import { bffApi } from '../../../testing/test-utils';
+import { bffApi } from '../../../testing/utils';
 import { AppRoutes } from '../../../universal/config/routes';
 import { AppState } from '../../../universal/types';
 import { appStateAtom } from '../../hooks/useAppState';
