+++ src/client/pages/Afis/AfisBetaalVoorkeuren.test.tsx
@@ -4,7 +4,7 @@ import { generatePath } from 'react-router-dom';
 import { MutableSnapshot } from 'recoil';
 
 import { AfisBusinessPartnerDetailsTransformed } from '../../../server/services/afis/afis-types';
-import { bffApi } from '../../../testing/test-utils';
+import { bffApi } from '../../../testing/utils';
 import { AppRoutes } from '../../../universal/config/routes';
 import { AppState } from '../../../universal/types';
 import { appStateAtom } from '../../hooks/useAppState';
