+++ src/client/pages/Search/Search.test.tsx
@@ -4,7 +4,7 @@ import { MutableSnapshot } from 'recoil';
 import { beforeAll, describe, it } from 'vitest';
 
 import Search from './Search';
-import { bffApi } from '../../../testing/test-utils';
+import { bffApi } from '../../../testing/utils';
 import { AppRoutes } from '../../../universal/config/routes';
 import { appStateAtom } from '../../hooks/useAppState';
 import MockApp from '../../pages/MockApp';
