+++ src/server/auth/auth-helpers.test.ts
@@ -22,7 +22,7 @@ import {
   getReqMockWithOidc,
   RequestMock,
   ResponseMock,
-} from '../../testing/test-utils';
+} from '../../testing/utils';
 import * as blacklist from '../services/session-blacklist';
 
 describe('auth-helpers', () => {
