+++ src/server/services/avg/avg.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, beforeEach, afterEach, vi, it } from 'vitest';
 import { fetchAVG, fetchAVGNotifications, transformAVGResponse } from './avg';
 import avgThemasResponse from '../../../../mocks/fixtures/avg-themas.json';
 import apiResponse from '../../../../mocks/fixtures/avg.json';
-import { remoteApi } from '../../../testing/test-utils';
+import { remoteApi } from '../../../testing/utils';
 import { AuthProfileAndToken } from '../../auth/auth-types';
 
 describe('AVG', () => {
