+++ src/server/services/bezwaren/bezwaren.test.ts
@@ -11,7 +11,7 @@ import bezwarenDocumenten from '../../../../mocks/fixtures/bezwaren-documents.js
 import bezwarenStatus from '../../../../mocks/fixtures/bezwaren-status.json';
 import bezwarenApiResponse from '../../../../mocks/fixtures/bezwaren.json';
 import { remoteApiHost } from '../../../testing/setup';
-import { remoteApi } from '../../../testing/test-utils';
+import { remoteApi } from '../../../testing/utils';
 import { range } from '../../../universal/helpers/utils';
 import { AuthProfileAndToken } from '../../auth/auth-types';
