+++ src/server/services/hli/stadspas.test.ts
@@ -8,7 +8,7 @@ import {
   StadspasDiscountTransactions,
   StadspasDiscountTransactionsResponseSource,
 } from './stadspas-types';
-import { remoteApi } from '../../../testing/test-utils';
+import { remoteApi } from '../../../testing/utils';
 import { AuthProfileAndToken } from '../../auth/auth-types';
 import * as encryptDecrypt from '../../helpers/encrypt-decrypt';
