+++ src/server/services/horeca.test.ts
@@ -3,7 +3,7 @@ import { afterAll, beforeAll, describe, expect, it } from 'vitest';
 
 import { fetchHorecaNotifications, fetchHorecaVergunningen } from './horeca';
 import vergunningenMockData from '../../../mocks/fixtures/vergunningen.json';
-import { remoteApi } from '../../testing/test-utils';
+import { remoteApi } from '../../testing/utils';
 import { AuthProfileAndToken } from '../auth/auth-types';
 
 describe('Horeca service', () => {
