+++ src/server/services/simple-connect/erfpacht.test.ts
@@ -7,7 +7,7 @@ import {
   fetchErfpachtNotifications,
   getConfigMain,
 } from './erfpacht';
-import { remoteApi } from '../../../testing/test-utils';
+import { remoteApi } from '../../../testing/utils';
 import { AuthProfileAndToken } from '../../auth/auth-types';
 
 describe('simple-connect/erfpacht', () => {
