+++ src/server/services/simple-connect/subsidie.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test } from 'vitest';
 
 import { fetchSubsidieNotifications } from './subsidie';
-import { remoteApi } from '../../../testing/test-utils';
+import { remoteApi } from '../../../testing/utils';
 import { Themas } from '../../../universal/config/thema';
 import { AuthProfileAndToken } from '../../auth/auth-types';
