+++ src/server/services/vergunningen-v2/decos-service.test.ts
@@ -12,7 +12,7 @@ import {
   fetchDecosZakenFromSource,
   forTesting,
 } from './decos-service';
-import { remoteApi } from '../../../testing/test-utils';
+import { remoteApi } from '../../../testing/utils';
 import { jsonCopy, range } from '../../../universal/helpers/utils';
 import { AuthProfileAndToken } from '../../auth/auth-types';
