+++ src/server/services/wmo/wmo-zorgned-service.test.ts
@@ -3,7 +3,7 @@ import Mockdate from 'mockdate';
 import { fetchZorgnedAanvragenWMO, forTesting } from './wmo-zorgned-service';
 import { fetchZorgnedAanvragenWMO, forTesting } from './wmo-zorgned-service';
 import { remoteApiHost } from '../../../testing/setup';
-import { remoteApi } from '../../../testing/test-utils';
+import { remoteApi } from '../../../testing/utils';
 import { AuthProfileAndToken } from '../../auth/auth-types';
 import * as request from '../../helpers/source-api-request';
 import {
